chore(fixture): revert change in fixture for LWE/GLWE prototypes transformation
