+++ concrete-core-fixture/src/fixture/lwe_ciphertext_vector_discarding_bootstrap_1.rs
@@ -3,7 +3,6 @@ use crate::fixture::Fixture;
 use crate::generation::prototyping::{
     PrototypesGlweCiphertextVector, PrototypesGlweSecretKey, PrototypesLweBootstrapKey,
     PrototypesLweCiphertextVector, PrototypesLweSecretKey, PrototypesPlaintextVector,
-    TransformsGlweToLweSecretKeyPrototype,
 };
 use crate::generation::synthesizing::{
     SynthesizesGlweCiphertextVector, SynthesizesLweBootstrapKey, SynthesizesLweCiphertextVector,
@@ -70,8 +69,7 @@ where
     OutputCiphertextVector: LweCiphertextVectorEntity,
     AccumulatorVector: GlweCiphertextVectorEntity,
     BootstrapKey: LweBootstrapKeyEntity,
-    Maker: TransformsGlweToLweSecretKeyPrototype<Precision, OutputKeyDistribution>
-        + SynthesizesLweBootstrapKey<
+    Maker: SynthesizesLweBootstrapKey<
             Precision,
             InputKeyDistribution,
             OutputKeyDistribution,
