+++ concrete-core-fixture/src/generation/prototyping/glwe_secret_key.rs
@@ -9,13 +9,14 @@ use crate::generation::{
 use concrete_commons::parameters::{GlweDimension, PolynomialSize};
 use concrete_core::prelude::{
     GlweSecretKeyGenerationEngine, GlweToLweSecretKeyTransformationEngine,
+    LweToGlweSecretKeyTransformationEngine,
 };
 
 /// A trait allowing to manipulate GLWE secret key prototypes.
 pub trait PrototypesGlweSecretKey<
     Precision: IntegerPrecision,
     KeyDistribution: KeyDistributionMarker,
->
+>: PrototypesLweSecretKey<Precision, KeyDistribution>
 {
     type GlweSecretKeyProto: GlweSecretKeyPrototype<
         Precision = Precision,
@@ -24,6 +25,17 @@ pub trait PrototypesGlweSecretKey<
     fn new_glwe_secret_key(
         &mut self,
         glwe_dimension: GlweDimension,
+        polynomial_size: PolynomialSize,
+    ) -> Self::GlweSecretKeyProto;
+    // Due to issues with cyclic imports, transformations between LWE and GLWE secret keys are made
+    // available here.
+    fn transform_glwe_secret_key_to_lwe_secret_key(
+        &mut self,
+        glwe_key: &Self::GlweSecretKeyProto,
+    ) -> Self::LweSecretKeyProto;
+    fn transform_lwe_secret_key_to_glwe_secret_key(
+        &mut self,
+        lwe_key: &Self::LweSecretKeyProto,
         polynomial_size: PolynomialSize,
     ) -> Self::GlweSecretKeyProto;
 }
@@ -39,6 +51,29 @@ impl PrototypesGlweSecretKey<Precision32, BinaryKeyDistribution> for Maker {
         ProtoBinaryGlweSecretKey32(
             self.default_engine
                 .generate_new_glwe_secret_key(glwe_dimension, polynomial_size)
+                .unwrap(),
+        )
+    }
+
+    fn transform_glwe_secret_key_to_lwe_secret_key(
+        &mut self,
+        glwe_key: &Self::GlweSecretKeyProto,
+    ) -> Self::LweSecretKeyProto {
+        ProtoBinaryLweSecretKey32(
+            self.default_engine
+                .transform_glwe_secret_key_to_lwe_secret_key(glwe_key.0.to_owned())
+                .unwrap(),
+        )
+    }
+
+    fn transform_lwe_secret_key_to_glwe_secret_key(
+        &mut self,
+        lwe_key: &Self::LweSecretKeyProto,
+        polynomial_size: PolynomialSize,
+    ) -> Self::GlweSecretKeyProto {
+        ProtoBinaryGlweSecretKey32(
+            self.default_engine
+                .transform_lwe_secret_key_to_glwe_secret_key(lwe_key.0.to_owned(), polynomial_size)
                 .unwrap(),
         )
     }
@@ -58,36 +93,7 @@ impl PrototypesGlweSecretKey<Precision64, BinaryKeyDistribution> for Maker {
                 .unwrap(),
         )
     }
-}
 
-/// A trait allowing to transform GLWE secret key prototypes to LWE secret key prototypes.
-pub trait TransformsGlweToLweSecretKeyPrototype<
-    Precision: IntegerPrecision,
-    KeyDistribution: KeyDistributionMarker,
->:
-    PrototypesGlweSecretKey<Precision, KeyDistribution>
-    + PrototypesLweSecretKey<Precision, KeyDistribution>
-{
-    fn transform_glwe_secret_key_to_lwe_secret_key(
-        &mut self,
-        glwe_key: &Self::GlweSecretKeyProto,
-    ) -> Self::LweSecretKeyProto;
-}
-
-impl TransformsGlweToLweSecretKeyPrototype<Precision32, BinaryKeyDistribution> for Maker {
-    fn transform_glwe_secret_key_to_lwe_secret_key(
-        &mut self,
-        glwe_key: &Self::GlweSecretKeyProto,
-    ) -> Self::LweSecretKeyProto {
-        ProtoBinaryLweSecretKey32(
-            self.default_engine
-                .transform_glwe_secret_key_to_lwe_secret_key(glwe_key.0.to_owned())
-                .unwrap(),
-        )
-    }
-}
-
-impl TransformsGlweToLweSecretKeyPrototype<Precision64, BinaryKeyDistribution> for Maker {
     fn transform_glwe_secret_key_to_lwe_secret_key(
         &mut self,
         glwe_key: &Self::GlweSecretKeyProto,
@@ -95,6 +101,18 @@ impl TransformsGlweToLweSecretKeyPrototype<Precision64, BinaryKeyDistribution> f
         ProtoBinaryLweSecretKey64(
             self.default_engine
                 .transform_glwe_secret_key_to_lwe_secret_key(glwe_key.0.to_owned())
+                .unwrap(),
+        )
+    }
+
+    fn transform_lwe_secret_key_to_glwe_secret_key(
+        &mut self,
+        lwe_key: &Self::LweSecretKeyProto,
+        polynomial_size: PolynomialSize,
+    ) -> Self::GlweSecretKeyProto {
+        ProtoBinaryGlweSecretKey64(
+            self.default_engine
+                .transform_lwe_secret_key_to_glwe_secret_key(lwe_key.0.to_owned(), polynomial_size)
                 .unwrap(),
         )
     }
